fix: Only remove AFK if user is AFK

+++ Hazel/plugins/afk.py
@@ -13,7 +13,7 @@ async def set_afk(_, m):
 async def handle_afk(c, m):
   uid = c.me.id
   is_afk = await db.is_afk(uid)
-  if uid == m.from_user.id:
+  if is_afk and uid == m.from_user.id:
     await db.remove_afk(uid)
     return await m.reply("Removed from afk.")
   elif is_afk:
